remove old files and transfer unmerged functionality
