--- Assets/Scripts/BinarySpacePartitioner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6410a303fe6020577944e474eeb35363
