--- Assets/Scripts/CorridorNode.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f5c35e6d21d65b3bf8deae6877c5f991
