--- Assets/Scripts/CorridorsGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fdc9766ecf0a2aef2b8a1ff2fd53e6cc
