--- Assets/Scripts/DugeonGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9e9c536139e2b7df6bdbd6f594bb012e
