--- Assets/Scripts/DungeonCreator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6795d3382dceb1a9aa269697de515a95
