+++ Assets/Scripts/DungeonGeneration/CorridorNode.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: f5c35e6d21d65b3bf8deae6877c5f991
+guid: 37688eac04fa68064ad9595b08ce8f99
