+++ Assets/Scripts/DungeonGeneration/CorridorsGenerator.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: fdc9766ecf0a2aef2b8a1ff2fd53e6cc
+guid: 2da3da718cb64024192a16b5119d3a50
