+++ Assets/Scripts/DungeonGeneration/DungeonCreator.cs
@@ -23,7 +23,7 @@ public class DungeonCreator : MonoBehaviour
     [Range(0, 2)]
     public int roomOffset;
 
-    public GameObject wallPrefab, pillarPrefab, playerPrefab, chestPrefab, enemyPrefab, shopPrefab, navPointPrefab;
+    public GameObject wallPrefab, pillarPrefab, playerPrefab, chestPrefab, enemyPrefab, shopPrefab, trapDoorPrefab, navPointPrefab;
     List<Vector3Int> possibleVerticalDoorPosition;
     List<Vector3Int> possibleHorizontalDoorPosition;
     List<Vector3Int> possibleHorizontalWallPosition;
@@ -62,6 +62,7 @@ public class DungeonCreator : MonoBehaviour
         possibleHorizontalWallPosition = new List<Vector3Int>();
         possibleVerticalWallPosition = new List<Vector3Int>();
         CreatePlayer(listOfRooms);
+        CreateTrapDoor(listOfRooms);
         for (int i = 0; i < listOfRooms.Count; i++)
         {
             CreateMesh(listOfRooms[i].BottomLeftAreaCorner, listOfRooms[i].TopRightAreaCorner, listOfRooms[i].Type);
@@ -176,12 +177,12 @@ public class DungeonCreator : MonoBehaviour
     }
 
 
-    private void SetRandomShopItems(ItemSlot[] slots) {
-        foreach (var slot in slots) {
+    private void SetRandomShopItems(ItemContainer items) {
+        foreach (ItemSlot slot in items.slots) {
             slot.storedItem = null;
 
             // select random definition
-            var random = UnityEngine.Random.Range(0f,1f);
+            float random = UnityEngine.Random.Range(0f,1f);
             foreach (ItemDefinition def in itemDefinitions.definitions) {
                 if (random <= def.shopProbability) {
                     slot.storedItem = def;
@@ -223,13 +224,45 @@ public class DungeonCreator : MonoBehaviour
                     shopY);
                 if(UnityEngine.Random.Range(0f,1f) < shopProb)
                 {
-                    var shop = Instantiate(shopPrefab, shopPos, Quaternion.identity);
-                    // TODO
-                    shop.GetComponent<ShopRenderer>().AddRandomItems();
+                    GameObject shop = Instantiate(shopPrefab, shopPos, Quaternion.identity);
+
+                    // compute random inventory of shop
+                    ItemContainer items = new();
+                    items.Resize(3);
+                    SetRandomShopItems(items);
+
+                    shop.GetComponent<ShopRenderer>().SetItems(items);
                 }
             }
         }
     }
+
+    // places the trapdoor in the room the furthest away from player
+    private void CreateTrapDoor(List<Node> listOfRooms) {
+        Transform player = GameObject.FindGameObjectWithTag("Player").transform;
+
+        float maxDist = 0f;
+        Node selectedRoom = null;
+        foreach (var room in listOfRooms) {
+            if (room.Type != "room") continue;
+
+            Vector3 currentPos = new Vector3(  (room.BottomLeftAreaCorner.x + room.BottomRightAreaCorner.x) / 2,
+                                        0,
+                                        (room.BottomLeftAreaCorner.y + room.TopLeftAreaCorner.y) / 2);
+            float dist = (player.position - currentPos).magnitude;
+            if (dist > maxDist) {
+                maxDist = dist;
+                selectedRoom = room;
+            }
+        }
+
+        // place in center of the selected room
+        Vector3 pos = new Vector3(  (selectedRoom.BottomLeftAreaCorner.x + selectedRoom.BottomRightAreaCorner.x) / 2,
+                                    0,
+                                    (selectedRoom.BottomLeftAreaCorner.y + selectedRoom.TopLeftAreaCorner.y) / 2);
+        Instantiate(trapDoorPrefab, pos, Quaternion.identity);
+    }
+
 
     private void CreateLoot(List<Node> listOfRooms)
     {
