+++ Assets/Scripts/DungeonGeneration/StructureHelper.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 40e7723f9d0f982afa3f1e79c4c26e53
+guid: 6c276668a756d0bfaaab7bd656c758f7
