--- Assets/Scripts/Line.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 36c52da5c1ee9fe78bc6b4c4b4c56eb2
