--- Assets/Scripts/Node.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 01c2f13e167b72251b938154d11350c5
