--- Assets/Scripts/RoomGenerator.cs
@@ -1,37 +1,0 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
-
-public class RoomGenerator
-{
-    private int maxIterations;
-    private int roomLengthMin;
-    private int roomWidthMin;
-
-    public RoomGenerator(int maxIterations, int roomLengthMin, int roomWidthMin)
-    {
-        this.maxIterations = maxIterations;
-        this.roomLengthMin = roomLengthMin;
-        this.roomWidthMin = roomWidthMin;
-    }
-
-    public List<RoomNode> GenerateRoomsInGivenSpaces(List<Node> roomSpaces, float roomBottomCornerModifier, float roomTopCornerMidifier, int roomOffset)
-    {
-        List<RoomNode> listToReturn = new List<RoomNode>();
-        foreach (var space in roomSpaces)
-        {
-            Vector2Int newBottomLeftPoint = StructureHelper.GenerateBottomLeftCornerBetween(
-                space.BottomLeftAreaCorner, space.TopRightAreaCorner, roomBottomCornerModifier, roomOffset);
-
-            Vector2Int newTopRightPoint = StructureHelper.GenerateTopRightCornerBetween(
-                space.BottomLeftAreaCorner, space.TopRightAreaCorner, roomTopCornerMidifier, roomOffset);
-            space.BottomLeftAreaCorner = newBottomLeftPoint;
-            space.TopRightAreaCorner = newTopRightPoint;
-            space.BottomRightAreaCorner = new Vector2Int(newTopRightPoint.x, newBottomLeftPoint.y);
-            space.TopLeftAreaCorner = new Vector2Int(newBottomLeftPoint.x, newTopRightPoint.y);
-            space.Type = "room";
-            listToReturn.Add((RoomNode)space);
-                
-        }
-        return listToReturn;
-    }
-}
