--- Assets/Scripts/RoomGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 129e0664332c3992181a0cdad30e838d
