--- Assets/Scripts/RoomNode.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: af3305999b8f56e2d9e9706886fb1299
