--- Assets/Scripts/StructureHelper.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 40e7723f9d0f982afa3f1e79c4c26e53
